requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
packaging==21.3
Pillow==9.1.1
py-cord==2.0.0rc1
pyparsing==3.0.9
pytesseract==0.3.9
python-dotenv==0.20.0
requests==2.28.0
urllib3==1.26.9
yarl==1.7.2
